requirements.txt
# Requirements para Sistema de Auditoria Fiscal ICMS v22.0 - Fase 6
# Base de dependências para execução 100% local com PostgreSQL + Agentes Reais

# Framework Web e API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Banco de Dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Agentes e LLM
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20
ollama==0.1.7

# Embeddings e RAG
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.18

# Base de Conhecimento Grafo
neo4j==5.15.0
py2neo==2021.2.4

# Processamento de Dados
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0

# Utilitários de ML/AI
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.0

# Validação e Avaliação
ragas==0.1.0
datasets==2.15.0

# Configuração e Logs
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# Autenticação e Segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilitários Gerais
requests==2.31.0
aiofiles==23.2.1
python-dateutil==2.8.2
tqdm==4.66.1

# Desenvolvimento e Testes
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.13.2
flake8==6.1.0

# Docker e Deploy
gunicorn==21.2.0
redis==5.0.1

# Fase 6 - Novas dependências
# Processamento de dados avançado
regex==2023.10.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Análise e estatísticas
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# Monitoramento e métricas
psutil==5.9.6
memory-profiler==0.61.0

# Validação de dados
jsonschema==4.20.0
cerberus==1.3.4

# Frontend Build (Node.js será necessário separadamente)
# Para o React frontend, será necessário:
# - Node.js 18+
# - npm ou yarn

# Integração futura com sistemas de estoque
# (Dependências serão adicionadas na v2.0)
xmltodict==0.13.0    # Para processamento XML de sistemas ERP
python-magic==0.4.27 # Para detecção de tipos de arquivo
chardet==5.2.0       # Para detecção de encoding

# Documentação e relatórios
reportlab==4.0.7     # Para geração de PDFs
jinja2==3.1.2        # Para templates de relatórios
